normalize Ray::dir on construction
Since ray construction now includes a fwd bump, it's important
that the bump be consistent in terms of distance.  With non-normalized
ray direction, the bump is a function of direction scale, which
is unappealing.
This also reduces the number of places I have to normalize the direction
in other code.
I'm not sure which is better, but unless normalizing ray direction is
prohibitively expensive, it's nice to be able to count on it being
normalized.